pyproject.toml
[tool.poetry]
name = "gadme"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
scipy = "^1.11.3"
ffprobe = "^0.5"
# --------- pytorch --------- #
torch = [
  {version = "^2.0.1+cu117", source = "pytorch", markers = "sys_platform == 'linux'"},
  {version = "^2.0.1", source = "PyPI", markers = "sys_platform == 'darwin'"}
]
torchvision = [
  {version = "0.15.2+cu117", source = "pytorch", markers = "sys_platform == 'linux'"},
  {version = "^0.15.2", source = "PyPI", markers = "sys_platform == 'darwin'"}
]

lightning= "^2.0"
torchmetrics= "^0.11"

# --------- hydra --------- #
hydra-core = "^1.3"
hydra-colorlog = "^1.2"
hydra-optuna-sweeper = "^1.2"

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
# rootutils       # standardizing the project root setup
# pre-commit      # hooks for applying linters on commit
# rich            # beautiful text formatting in terminal
# pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
rootutils = "^1.0.7"
pre-commit = "^3.5.0"
rich = "^13.6.0"
pytest = "^7.4.2"
sh = "^2.0.6"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
tensorboard = "^2.14.1"
ipywidgets = "^8.1.1"
jupyter = "^1.0.0"
datasets = "^2.14.6"
audiomentations = "^0.33.0"
torch-audiomentations = "^0.11.0"
wandb = "^0.16.0"
transformers = "^4.35.0"
pyrootutils = "^1.0.4"
mlflow = "^2.11.0"
pydub = "^0.25.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu117"
priority = "explicit"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
